Use the delete modal for deleting activities.
