+++ spec/features/aeon_requests_spec.rb
@@ -176,6 +176,7 @@ RSpec.describe 'Requests', :js do
       expect(page).to have_text('An Activity Book')
 
       click_on 'Remove An Activity Book from activity'
+      click_on 'Yes - Delete'
 
       expect(page).to have_no_text('An Activity Book')
     end
@@ -200,6 +201,7 @@ RSpec.describe 'Requests', :js do
         expect(page).to have_text('Another Activity Book')
 
         click_on 'Remove An Activity Book from activity'
+        click_on 'Yes - Delete'
 
         expect(page).to have_no_text('An Activity Book')
         expect(page).to have_text('Another Activity Book')
@@ -239,6 +241,7 @@ RSpec.describe 'Requests', :js do
         expect(page).to have_text('An Activity Book')
 
         click_on 'Remove An Activity Book from activity'
+        click_on 'Yes - Delete'
 
         expect(page).to have_no_text('An Activity Book')
       end
